[Fix] rename the key, fix utils import path
