+++ src/components/organisms/Map.jsx
@@ -2,7 +2,7 @@ import GoogleMapReact from "google-map-react";
 
 import styled from "styled-components";
 
-import MAP_KEY from "@utils/utils";
+import { GOOGLE_MAP_API_KEY } from "@utils/utils";
 
 const AnyReactComponent = ({ text }) => <div>{text}</div>;
 
@@ -18,7 +18,7 @@ const Map = ({ lat, lng }) => {
   return (
     <StyledMap>
       <GoogleMapReact
-        bootstrapURLKeys={{ key: MAP_KEY }}
+        bootstrapURLKeys={{ key: GOOGLE_MAP_API_KEY }}
         defaultCenter={defaultProps.center}
         defaultZoom={defaultProps.zoom}
       >
