+++ src/utils/utils.jsx
@@ -1,5 +1,7 @@
 export const BASE_URL = import.meta.env.VITE_APP_BASE_URL;
+
 export const GITHUB_URL = "https://github.com/GDSC-KNU/3rd-sc-5-PIKPET-FE";
 export const NOTION_URL =
   "https://pikpet.notion.site/Home-374c20373355402f9c6ff026cfd0c360?pvs=4";
-export const MAP_KEY = import.meta.env.GOOGLE_MAP_API_KEY;
+
+export const GOOGLE_MAP_API_KEY = import.meta.env.VITE_APP_GOOGLE_MAP_API_KEY;
